requirements.txt
ag2[openai]
Pillow
numpy
python-dotenv
